system/converter/requirements.txt
python-pptx>=0.6.21
python-docx>=1.1.0
openpyxl>=3.1.2
reportlab>=4.1.0
python-frontmatter>=1.1.0
Markdown>=3.5.0
Pillow>=10.0.0
